gscholar_scraper/requirements.txt
# Scrapy framework
scrapy>=1.0
stem
python-dotenv
scrapy-fake-useragent
sqlalchemy
psycopg2
scrapyd-client
